Fix bug archimate reader - add missing dummy reader argument

+++ src/pyArchimate/writers/archiWriter.py
@@ -165,7 +165,7 @@ def archi_writer(model: Model, file_path: str):
                 'id': node.uuid,
             })
             if node.font_name is not None and node.font_size is not None:
-                size = str(int(node.font_size))
+                size = str(float(node.font_size))
                 child.set('font', f"1|{node.font_name}|{size}.0|0|WINDOWS|1|0|0|0|0|0|0|0|0|1|0|0|0|0|{node.font_name}")
             if node.font_color is not None:
                 child.set('fontColor', node.font_color.lower())
